requirements.txt
requests
beautifulsoup4
lxml
pycryptodome
icalendar
pytz
ddddocr
PyYAML
python-dotenv